refactor(simulation): move detail use case into feature module
