+++ src/main/java/com/renewsim/backend/simulation_service/detail/web/SimulationDetailController.java
@@ -1,6 +1,6 @@
 package com.renewsim.backend.simulation_service.detail.web;
 
-import com.renewsim.backend.simulation_service.application.detailSimulation.GetRealSimulationUseCase;
+import com.renewsim.backend.simulation_service.detail.application.GetRealSimulationUseCase;
 import com.renewsim.backend.simulation_service.shared.web.SimulationRequestContext;
 import com.renewsim.backend.simulation_service.shared.web.SimulationDetailsWebMapper;
 import com.renewsim.backend.simulation_service.shared.web.SimulationRequestContextFactory;
